+++ src/tests/test_tenant_resources.py
@@ -7,7 +7,6 @@ resources = [
   "asg",
   "cronjob", 
   "job",
-  "secret",
   "configmap",
   "lambda",
   "rds",